Part out remaining 2 methods WOS Contributions module
Each method was used in exactly one place. Ported relevant tests over
also.
