+++ lib/notification_manager.rb
@@ -17,7 +17,7 @@ class NotificationManager
         log_exception(pubmed_logger, log_message, e)
       when Cap::AuthorsPoller, Cap::Client
         log_exception(cap_logger, log_message, e)
-      when WebOfScience::Client, WebOfScience::Contributions,
+      when WebOfScience::Client,
         WebOfScience::Harvester, WebOfScience::ProcessRecords, WebOfScience::Record
         log_exception(WebOfScience.logger, log_message, e)
       else
