+++ lib/web_of_science/harvester.rb
@@ -1,12 +1,9 @@
 module WebOfScience
-
   # Application logic to harvest publications from Web of Science;
   # This is the bridge between the WebOfScience API and the SUL-PUB application.
   # This class is responsible for processing WebOfScience API response data
   # to integrate it into the application data models.
   class Harvester < ::Harvester::Base
-    include WebOfScience::Contributions
-
     # @param [Enumerable<Author>] authors
     # @param [Hash] options
     # @return [void]
@@ -55,6 +52,25 @@ module WebOfScience
     private
 
       delegate :logger, :queries, to: :WebOfScience
+
+      # Find any matching contributions by author and WOS-UID; create a contribution for any
+      # existing publication without one for the author in question.
+      # @param author [Author]
+      # @param uids [Array<String>]
+      # @return [Array<String>] uids guaranteed to have matching Publication and Contribution
+      def author_contributions(author, uids)
+        Publication.where(wos_uid: uids).find_each.map do |pub|
+          pub.contributions.find_or_create_by!(author_id: author.id) do |contrib|
+            contrib.assign_attributes(
+              cap_profile_id: author.cap_profile_id,
+              featured: false, status: 'new', visibility: 'private'
+            )
+            pub.pubhash_needs_update! # Add to pub_hash[:authorship]
+            pub.save! # contrib.save! not needed
+          end
+          pub.wos_uid
+        end
+      end
 
       # Consistent log prefix for status updates
       # @param author [Author]
