+++ lib/web_of_science/process_records.rb
@@ -4,7 +4,6 @@ module WebOfScience
   # Process records retrieved by any means; this is a progressive filtering of the harvested records to identify
   # those records that should create a new Publication.pub_hash, PublicationIdentifier(s) and Contribution(s).
   class ProcessRecords
-    include WebOfScience::Contributions
     include WebOfScience::ProcessPubmed
 
     # @param author [Author]
@@ -100,6 +99,29 @@ module WebOfScience
         links_client.links records.map { |rec| rec.uid if rec.database == 'WOS' }.compact
       rescue StandardError => err
         NotificationManager.error(err, "Author: #{author.id}, retrieve_links failed", self)
+      end
+
+      # Does record have a contribution for this author? (based on matching PublicationIdentifiers)
+      # Note: must use unique identifiers, don't use ISSN or similar series level identifiers
+      # We search for all PubIDs at once instead of serial queries.  No need to hit the same table multiple times.
+      # @param [WebOfScience::Record] record
+      # @return [::Publication, nil] a matched or newly minted Contribution
+      def matching_publication(record)
+        Publication.joins(:publication_identifiers).where(
+          "publication_identifiers.identifier_value IS NOT NULL AND (
+           (publication_identifiers.identifier_type = 'WosUID' AND publication_identifiers.identifier_value = ?) OR
+           (publication_identifiers.identifier_type = 'WosItemID' AND publication_identifiers.identifier_value = ?) OR
+           (publication_identifiers.identifier_type = 'doi' AND publication_identifiers.identifier_value = ?) OR
+           (publication_identifiers.identifier_type = 'pmid' AND publication_identifiers.identifier_value = ?))",
+           record.uid, record.wos_item_id, record.doi, record.pmid
+        ).order(
+          "CASE
+            WHEN publication_identifiers.identifier_type = 'WosUID' THEN 0
+            WHEN publication_identifiers.identifier_type = 'WosItemID' THEN 1
+            WHEN publication_identifiers.identifier_type = 'doi' THEN 2
+            WHEN publication_identifiers.identifier_type = 'pmid' THEN 3
+           END"
+        ).first
       end
   end
 end
