+++ spec/spec_helper.rb
@@ -84,6 +84,7 @@ RSpec.configure do |config|
 end
 
 RSpec::Matchers.define_negated_matcher :exclude, :include
+RSpec::Matchers.define_negated_matcher :not_change, :change
 
 require 'vcr'
 VCR.configure do |c|
